fix: 修复 clippy -D warnings 门禁失败 (collapsible_if/too_many_arguments 等)
CI 的 clippy 门禁 (cargo clippy -- -D warnings) 此前从未通过，一直被
tx-di-macros 的 4 处 collapsible_if 阻断。本次全面修复所有 pre-existing
clippy 告警，使门禁得以通过：

- tx-di-macros: collapsible_if 折叠为 let-chains (4处)
- tx-di-core: mem_replace_with_default / derivable_impls / type_complexity / needless_borrow
- 插件 crate (tx_di_log/toasty/cache/file/nacos/axum/job): collapsible_if、
  type_complexity、redundant_closure、unnecessary_cast、derivable_impls、
  io_other_error 等
- admin_domain/admin_app/admin_infra/admin_api: ok_or_else→ok_or、
  needless_return_with_question_mark、collapsible_if、too_many_arguments(加
  #[allow])、redundant import、empty_line_after_doc_comment 等

+++ examples/tx_admin/admin_app/src/job/dto.rs
@@ -9,7 +9,7 @@ pub use admin_proto::{
 /// 将 InfrustJob 转换为 proto JobResponse
 pub fn job_to_response(job: InfrustJob) -> JobResponse {
     JobResponse {
-        id: job.id as u64,
+        id: job.id,
         name: job.name,
         status: job.status as i32,
         handler_name: job.handler_name,
@@ -24,8 +24,8 @@ pub fn job_to_response(job: InfrustJob) -> JobResponse {
 /// 将 InfrustJobLog 转换为 proto JobLogResponse
 pub fn job_log_to_response(log: InfrustJobLog) -> JobLogResponse {
     JobLogResponse {
-        id: log.id as u64,
-        job_id: log.job_id as u64,
+        id: log.id,
+        job_id: log.job_id,
         handler_name: log.handler_name,
         handler_param: log.handler_param,
         execute_index: log.execute_index as i32,
